Bug#33141958 - THE FIRST ASAN UAF ISSUE OF MYSQL SERVER
Statements which commits a transaction implicitly or explicitly are
not allowed inside a trigger and a stored function. CREATE TRIGGER
and CREATE FUNCTION reports an error in this case.

DROP TABLESPACE statement implicitly commits a transaction. But
while creating a trigger and a stored function, this statement is
allowed by mistake. Because of which, assertion to make sure
statement is allowed inside a trigger or a stored function, fails
while executing it in the debug build. An error is reported in the
non-debug build.

To fix this issue, included DROP TABLESPACE statement in the
list of statements which are not allowed inside a trigger and a
stored function. CREATE TRIGGER and CREATE FUNCTION reports an
error if body contains DROP TABLESPACE statement.

Change-Id: I18b89ae4b63e7b1b40d5b3961a7b9eb9711f849c