Bug#32989716: MISSING CORRECT CLOSE SESSION HANDLING DURING               KEYRING MIGRATION
Description: Keyring migration server uses a mysql
connection to perform online migration. This mysql session
is not closed gracefully after the migration. This leads to
a 'Note' being printed on the online server error log:-
"[Note] Aborted connection x to db: 'xyz' user: 'abc' host:
'localhost' (Got an error reading communication packets)"

Analysis: After the online keyring migration, migration
server process is aborted before gracefully closing the
mysql session. This happens because the migration server
process exits before calling the destructor of
'Migrate_keyring', where the closing of the mysql session
is handled.

Fix: This issue is fixed in 8.0 as part of a code
refactoring done in this area by WL#13446. Back-porting the
relevant code changes to 5.7 to fix this bug and to make
the code consistent across versions.

RB#27186