PS-7883
This is improvement proposal.

The aim of this patch is to switch the value of session variable
rocksdb_write_disable_val to the value compatible with global variable
rocksdb_flush_log_at_trx_commit only for the time we detect that session
variable is not compatible with the global one.
Once the global variable becomes compatible with the old session
variable again, the session variable is set back to its original value.

Problem still not solved:
If we set global variable to the incompatible value and query for the
value of session variable, we will get its incompatible value.
Only after 1st query (eg insert) processed by session, the session
variable is temporarily set to the compatible value.
The same is observed for restoring the original session variable value
once the global variable switches back to the compatible one.

This problem is not easy to solve because there is no 'on_value_read'
hook for the variable, where we could calculate the current value of
session variable. MySql layer rather digs directly into thread's
variables hash and gets values from there, so we have no 'hook' where we
could put our logic.

Additionally extracted the common logic into
fix_write_disable_wal_value() function.