PS-8028: Server crashes with Assertion `m_prebuilt->table->n_ref_count > 0' failed
https://jira.percona.com/browse/PS-8028

Cause:
5.7 does not support atomic DDL. It may happen that MySql's frm file is
out of sync with InnodDB dictionary. In such a case the table that is
part of InnoDB partitioned table is not opened which causes asserts.

Fix:
We are not able to recover from this case automatically, but we can
improve the error handling.
If the partitioned table cannot be opened, and its 1st partition is
marked as corrupted, return error without proceeding with further
opening.

Also, logging was improved as it may happen that when the inconsistency
is detected (opening the table), partition names are not initialized
yet, which results with the spoiled log.