Bug#27492990 	PARALLEL REPLICATION ALWAYS FAILS WITH SPECIFIC WORKLOAD FROM SYSBENCH
Background

Change-Id: I74bc82fcdfaae28e2c005b46b93084b45653fbb8
----------
If we want to apply binary log files as relay log files in a server that has
the same server id as the id in the files then we can do it using
replicate_same_server_id flag.

Problem
-------
The replica applier coordinator fails to adapt its internal transaction
boundary trackers in the presence of a Rotate event when
replicate_same_server_id is enabled and the event has the same server id.

Solution
--------
In this case, we should make sure that the value of replicate_same_server_id
is properly checked in the conditions in
`Rotate_log_event::do_update_pos(Relay_log_info *rli)`.

Change-Id: I74bc82fcdfaae28e2c005b46b93084b45653fbb8